experimental: add transform-origin extractor for individual proeprty parsing
